fix: refactor Sb card to template
