+++ src/components/Card/SbCard.vue
@@ -1,0 +1,34 @@
+<template>
+  <div
+    class="sb-card"
+    :class="{
+      'sb-card--loading': isLoading,
+    }"
+  >
+    <slot></slot>
+  </div>
+</template>
+<script>
+export default {
+  name: 'SbCard',
+  provide() {
+    return {
+      cardContext: () => this.cardContext,
+    }
+  },
+  props: {
+    isLoading: {
+      type: Boolean,
+      default: false,
+    },
+  },
+  computed: {
+    cardContext() {
+      return {
+        // states
+        isLoading: this.isLoading,
+      }
+    },
+  },
+}
+</script>
