+++ src/components/Card/components/SbCardFooter.vue
@@ -1,0 +1,10 @@
+<template>
+  <div class="sb-card__footer">
+    <slot></slot>
+  </div>
+</template>
+<script>
+export default {
+  name: 'SbCardFooter',
+}
+</script>
