+++ src/components/Card/components/SbCardHeader.vue
@@ -1,0 +1,30 @@
+<template>
+  <div class="sb-card__header" v-bind="$attrs">
+    <component :is="as" v-if="title" class="sb-card__title"
+      >{{ title }}
+      <slot></slot>
+    </component>
+    <SbCardOptions v-if="options" :options="options"></SbCardOptions>
+  </div>
+</template>
+<script>
+import SbCardOptions from './SbCardOptions.vue'
+export default {
+  name: 'SbCardHeader',
+  components: { SbCardOptions },
+  props: {
+    as: {
+      type: String,
+      default: 'span',
+    },
+    options: {
+      type: Array,
+      default: null,
+    },
+    title: {
+      type: String,
+      default: null,
+    },
+  },
+}
+</script>
